בוט שרות וירטואלי PYTHON/requirements.txt.txt
streamlit
google-genai
matplotlib
numpy